chore: rename utils to other because utils doesn't mean anything
