+++ other/httpStatusCodes.ts
@@ -1,4 +1,4 @@
-export const STATUS_CODES = {
+export const HTTP_STATUS_CODES = {
   UNPROCESSABLE_ENTITY: 422,
   UNAUTHORIZED: 401,
 };
