+++ tests/utils/seed-test-users.ts
@@ -1,5 +1,5 @@
-import db from "../../utils/db";
-import createUser from "../../utils/createUser";
+import db from "../../other/db";
+import createUser from "../../other/createUser";
 
 const TEST_USERS = [
   { username: "test", password: "1234" },
